+++ parler/managers.py
@@ -37,6 +37,21 @@ class TranslatableQuerySet(QuerySet):
         return self
 
 
+    def translated(self, *language_codes):
+        """
+        Only return objects which are translated in the given languages.
+
+        NOTE: due to Django `ORM limitations <https://docs.djangoproject.com/en/dev/topics/db/queries/#spanning-multi-valued-relationships>`_,
+        this method can't be combined with other filters that access the translated fields.
+        """
+        relname = self.model._translations_field
+
+        if len(language_codes) == 1:
+            return self.filter(**{relname + '__language_code': language_codes[0]})
+        else:
+            return self.filter(**{relname + '__language_code__in': language_codes}).distinct()
+
+
     def iterator(self):
         """
         Overwritten iterator which will apply the decorate functions before returning it.
@@ -67,6 +82,15 @@ class TranslatableManager(models.Manager):
         Set the language code to assign to objects retrieved using this Manager.
         """
         return self.get_query_set().language(language_code)
+
+    def translated(self, *language_codes):
+        """
+        Only return objects which are translated in the given languages.
+
+        NOTE: due to Django `ORM limitations <https://docs.djangoproject.com/en/dev/topics/db/queries/#spanning-multi-valued-relationships>`_,
+        this method can't be combined with other filters that access the translated fields.
+        """
+        return self.get_query_set().translated(*language_codes)
 
 
 # Export the names in django-hvad style too: